Lasting critique, a first step into freedom.
A first step into pleasure, paradise, leisure, and abundance. A walk into the grand life of a future worth living, and the freedom we aspire to acquire. One day with all great effort of our people this will be peace and not agony and torment of the fighting of our enemies, and the slewing of our peaieydeiments.